fix: the issue of custom screensaver self-startup
 1、The cause is that the dbus interface is not masked.
 2、Rearrange the logic that show only one Settings window

Log: fix the issue of custom screensaver self-startup
Bug: https://pms.uniontech.com/bug-view-177209.html
